+++ built_stream/lib/stream_types.dart
@@ -53,12 +53,12 @@ class StreamError implements StreamState {
       .._payload = streamError.payload;
   }
 
-  Error _error;
+  dynamic _error;
   dynamic _payload;
   List<ErrorLocation> _locations;
 
   dynamic get payload => _payload;
-  Error get error => _error;
+  dynamic get error => _error;
   List<ErrorLocation> get locations => _locations;
 
   @override