+++ built_stream_generator/lib/src/property.dart
@@ -7,12 +7,15 @@
 /// print(property) // will show 'String username' to console
 /// ```
 class Property {
-  final String type;
-  final String name;
+  final String _type;
+  final String _name;
+  Property(this._type, this._name);
 
-  Property(this.type, this.name);
-  @override
-  String toString() {
-    return '$type $name';
-  }
+  String get privateName => '_$_name';
+  String get publicName => _name;
+
+  String get type => _type;
+
+  String get privateDeclaration => '$type _$_name';
+  String get publicDeclaration => '$type $_name';
 }